% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victory.R
\name{victory-shiny}
\alias{victory-shiny}
\alias{victoryOutput}
\alias{renderVictory}
\alias{victory_html}
\title{Shiny bindings for victory}
\usage{
victoryOutput(outputId, width = "100\%", height = "400px")

renderVictory(expr, env = parent.frame(), quoted = FALSE)

victory_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a victory}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using victory within Shiny
applications and interactive Rmd documents.
}
